refactor(headless-client): move IPC servers into their own module (#5360)
Makes #5357 easier by freeing up space in the catch-all `platform` mod
