+++ rust/gui-client/src-tauri/src/client/ipc/windows.rs
@@ -11,7 +11,7 @@ pub(crate) type IpcStream = named_pipe::NamedPipeClient;
 /// This is async on Linux
 #[allow(clippy::unused_async)]
 pub(crate) async fn connect_to_service() -> Result<IpcStream> {
-    let path = firezone_headless_client::windows::pipe_path();
+    let path = firezone_headless_client::ipc::platform::pipe_path();
     let stream = named_pipe::ClientOptions::new()
         .open(path)
         .with_context(|| "Couldn't connect to named pipe server at `{path}`")?;
