+++ rust/headless-client/src/ipc.rs
@@ -1,0 +1,11 @@
+#[cfg(target_os = "linux")]
+mod linux;
+#[cfg(target_os = "linux")]
+use linux as platform;
+
+#[cfg(target_os = "windows")]
+pub mod windows;
+#[cfg(target_os = "windows")]
+pub use windows as platform;
+
+pub(crate) use platform::{Server, Stream};
